Added getSample() to JointDistributionEmperical which allows one to retrieve a sample from the joint distribution.
Changed getPMF() in JointDistributionEmperical such that it returns an
the *unconditional* probability distribution for the desired
good whenever a conditional probability does not exist. Previously, it
returned an array of all zeros -- which did not make sense for a variety
of reasons, not the least of which was that an array of all zeros
violates the definition of a PDF (which must total 1).